FUR-2759 Support for >100 associations
This work was originally done for PHIS+, but was never pushed to openfurther
(due to various reasons).  This brings the last bit of code back into master
(for 1.4.4).

Add support for modifying the number of associations returned from DTS.  The
default is, and has always been, 100 (and remains so unless specified otherwise).